refactor(protocol-designer): refactor EditInstrumentsModal
Export most parts and make EditInstrumentsModal as the main component
